LPS-62907 rename package
