+++ modules/apps/product-navigation/product-navigation-control-menu-api/src/main/java/com/liferay/product/navigation/control/menu/ControlMenuEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.control.menu;
+package com.liferay.product.navigation.control.menu;
 
 import com.liferay.portal.kernel.exception.PortalException;
