+++ modules/apps/product-navigation/product-navigation-control-menu-api/src/main/java/com/liferay/product/navigation/control/menu/categories/UserControlMenuCategory.java
@@ -1,0 +1,47 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.product.navigation.control.menu.categories;
+
+import com.liferay.product.navigation.control.menu.ControlMenuCategory;
+import com.liferay.product.navigation.control.menu.constants.ControlMenuCategoryKeys;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Julio Camarero
+ */
+@Component(
+	immediate = true,
+	property = {
+		"control.menu.category.key=" + ControlMenuCategoryKeys.ROOT,
+		"service.ranking:Integer=300"
+	},
+	service = ControlMenuCategory.class
+)
+public class UserControlMenuCategory implements ControlMenuCategory {
+
+	@Override
+	public String getKey() {
+		return ControlMenuCategoryKeys.USER;
+	}
+
+	@Override
+	public boolean hasAccessPermission(HttpServletRequest request) {
+		return true;
+	}
+
+}
