+++ modules/apps/product-navigation/product-navigation-control-menu-api/src/main/java/com/liferay/product/navigation/control/menu/constants/ControlMenuCategoryKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.control.menu.constants;
+package com.liferay.product.navigation.control.menu.constants;
 
 /**
  * @author Julio Camarero
