+++ modules/apps/product-navigation/product-navigation-taglib/src/main/java/com/liferay/control/menu/taglib/servlet/taglib/ControlMenuTag.java
@@ -14,11 +14,11 @@
 
 package com.liferay.control.menu.taglib.servlet.taglib;
 
-import com.liferay.control.menu.ControlMenuCategory;
-import com.liferay.control.menu.constants.ControlMenuCategoryKeys;
 import com.liferay.control.menu.taglib.servlet.ServletContextUtil;
-import com.liferay.control.menu.util.ControlMenuCategoryRegistry;
-import com.liferay.control.menu.util.ControlMenuEntryRegistry;
+import com.liferay.product.navigation.control.menu.ControlMenuCategory;
+import com.liferay.product.navigation.control.menu.constants.ControlMenuCategoryKeys;
+import com.liferay.product.navigation.control.menu.util.ControlMenuCategoryRegistry;
+import com.liferay.product.navigation.control.menu.util.ControlMenuEntryRegistry;
 import com.liferay.taglib.util.IncludeTag;
 
 import java.util.List;
