+++ modules/apps/product-navigation/product-navigation-taglib/src/main/resources/META-INF/resources/init.jsp
@@ -23,9 +23,9 @@ taglib prefix="liferay-portlet" uri="http://liferay.com/tld/portlet" %><%@
 taglib prefix="liferay-theme" uri="http://liferay.com/tld/theme" %><%@
 taglib prefix="liferay-ui" uri="http://liferay.com/tld/ui" %>
 
-<%@ page import="com.liferay.control.menu.ControlMenuCategory" %><%@
-page import="com.liferay.control.menu.ControlMenuEntry" %><%@
-page import="com.liferay.control.menu.util.ControlMenuEntryRegistry" %><%@
+<%@ page import="com.liferay.product.navigation.control.menu.ControlMenuCategory" %><%@
+page import="com.liferay.product.navigation.control.menu.ControlMenuEntry" %><%@
+page import="com.liferay.product.navigation.control.menu.util.ControlMenuEntryRegistry" %><%@
 page import="com.liferay.taglib.servlet.PipingServletResponse" %>
 
 <%@ page import="java.util.List" %>
